Module ncollide2df64​::bounding_volume​

Bounding volumes.

Structs

AABB
An Axis Aligned Bounding Box.

BoundingSphere
A Bounding Sphere.

SpacializedCone
A normal cone with a bounding sphere.

Traits

BoundingVolume
Trait of bounding volumes.

HasAABB
Trait of objects that can be bounded by an AABB.

HasBoundingSphere
Trait implemented by objects having a bounding sphere.

HasBoundingVolume
Traits of objects having a bounding volume.

LooseBoundingVolume
Trait of loose bounding volume.

Functions

ball_aabb
Computes the Axis-Aligned Bounding Box of a ball.

implicit_shape_aabb
Computes the AABB of an implicit shape.

point_cloud_aabb
Computes the AABB of a set of point.

point_cloud_bounding_sphere
Computes a bounding sphere of the specified set of point.

point_cloud_bounding_sphere_with_center
Computes the bounding sphere of a set of point, given its center.