Trait ncollide2df64​::bounding_volume​::BoundingVolume​

pub trait BoundingVolume {
    fn intersects(&self, &Self) -> bool;
    fn contains(&self, &Self) -> bool;
    fn merge(&mut self, &Self);
    fn merged(&self, &Self) -> Self;
}

Trait of bounding volumes.

Bounding volumes are coarse approximations of geometric primitives. It usually have constant time intersection, inclusion test. Two bounding volume must also be mergeable into a bigger bounding volume.

Required Methods

fn intersects(&self, &Self) -> bool

Checks if this bounding volume intersect with another one.

fn contains(&self, &Self) -> bool

Checks if this bounding volume contains another one.

fn merge(&mut self, &Self)

Merges this bounding volume with another one. The merge is done in-place.

fn merged(&self, &Self) -> Self

Merges this bounding volume with another one.