Struct ncollide2df64​::broad​::DBVTBroadPhase​

pub struct DBVTBroadPhase<B, BV, D, DV> {
    // some fields omitted
}

Broad phase based on a Dynamic Bounding Volume Tree.

It uses two separate trees: one for static objects and which is never updated, and one for moving objects.

Methods

impl<B: 'static + HasBoundingVolume<BV> + Clone, BV: 'static + LooseBoundingVolume + Translation<Vect> + Clone, D: Dispatcher<B, B, DV>, DV> DBVTBroadPhase<B, BV, D, DV>

fn new(dispatcher: D, margin: Scalar) -> DBVTBroadPhase<B, BV, D, DV>

Creates a new broad phase based on a Dynamic Bounding Volume Tree.

fn num_interferences(&self) -> uint

Number of interferences detected by this broad phase.