Struct ncollide2df64​::geom​::AnnotatedMinkowskiSum​

pub struct AnnotatedMinkowskiSum<'a, G1, G2> {
    // some fields omitted
}

Same as the MinkowskiSum but with a support mapping which keeps track of the original supports points from the two wrapped geometries.

G1: type of the first object involved on the sum.

G2: type of the second object involved on the sum.

Methods

impl<'a, G1, G2> AnnotatedMinkowskiSum<'a, G1, G2>

fn new(m1: &'a Matrix, g1: &'a G1, m2: &'a Matrix, g2: &'a G2) -> AnnotatedMinkowskiSum<'a, G1, G2>

Builds the Minkowski sum of two geometries. Since the representation is implicit, this is done in constant time.

fn m1(&self) -> &'a Matrix

The transformation matrix of the first geometry of this Minkowski Sum.

fn m2(&self) -> &'a Matrix

The transformation matrix of the second geometry of this Minkowski Sum.

fn g1(&self) -> &'a G1

The first geometry of this Minkowski Sum.

fn g2(&self) -> &'a G2

The second geometry of this Minkowski Sum.