Struct ncollide2df64​::geom​::BezierSurface​

pub struct BezierSurface {
    // some fields omitted
}

Procedural generator of non-rational Bézier surfaces.

Methods

impl BezierSurface

fn new(control_points: Vec<Vect>, nupoints: uint, nvpoints: uint) -> BezierSurface

Creates a new procedural generator of non-rational bézier surfaces.

Parameters:

control_points: The control points of the bézier surface.

nupoints: The number of control points per u-iso-curve.

nvpoints: The number of control points per v-iso-curve.

Failures:

Fails if the vector of control points does not contain exactly nupoints * nvpoints elements.

fn new_with_degrees(degree_u: uint, degree_v: uint) -> BezierSurface

Creates a new bézier curve of degrees u_degree x v_degree.

Its control points are initialized to zero.

fn new_evaluation_cache() -> BezierSurfaceEvaluationCache

Creates a cache to avaluate a bezier surface.

fn reset_with_degrees(&mut self, degree_u: uint, degree_v: uint) -> bool

Changes the degree of this surface and reset its control points to zero.

If the surface already has the specified degrees, nothing is changed, and false is returned. Otherwise true is returned.

fn control_points<'a>(&'a self) -> &'a [Vect]

The control points of this bézier surface.

fn nupoints(&self) -> uint

The number of control points per u-iso-curve.

fn nvpoints(&self) -> uint

The number of control points per v-iso-curve.

fn degree_u(&self) -> uint

The degree of u-iso-curves.

fn degree_v(&self) -> uint

The degree of v-iso-curves.

fn slice_u<'a>(&'a self, i: uint) -> VecSlice<'a, Vect>

Get the i-th set of control points along the u parametric direction.

fn slice_v<'a>(&'a self, i: uint) -> VecSlice<'a, Vect>

Get the i-th set of control points along the v parametric direction.

fn mut_slice_u<'a>(&'a mut self, i: uint) -> VecSliceMut<'a, Vect>

Mutably get the i-th set of control points along the u parametric direction.

fn mut_slice_v<'a>(&'a mut self, i: uint) -> VecSliceMut<'a, Vect>

Mutably get the i-th set of control points along the v parametric direction.

impl BezierSurface

fn at(&self, u: &Scalar, v: &Scalar, cache: &mut BezierSurfaceEvaluationCache) -> Vect

Evaluates this bezier surface at the given parameter t.

fn subdivide_u(&self, u: &Scalar, out1: &mut BezierSurface, out2: &mut BezierSurface)

Subdivides this bezier at the given isocurve of parameter u.

The union of the two resulting surfaces equals the original one.

fn subdivide_v(&self, v: &Scalar, out1: &mut BezierSurface, out2: &mut BezierSurface)

Subdivides this bezier at the given isocurve of parameter v.

The union of the two resulting surfaces equals the original one.

fn diff_u(&self, out: &mut BezierSurface)

Computes the derivative wrt. u of this bezier surface.

fn diff_v(&self, out: &mut BezierSurface)

Computes the derivative wrt. v of this bezier surface.

fn endpoint_00<'a>(&'a self) -> &'a Vect

Gets the endpoint at parameters u = 0 and v = 0.

fn endpoint_01<'a>(&'a self) -> &'a Vect

Gets the endpoint at parameters u = 0 and v = 1.

fn endpoint_10<'a>(&'a self) -> &'a Vect

Gets the endpoint at parameters u = 0 and v = 0.

fn endpoint_11<'a>(&'a self) -> &'a Vect

Gets the endpoint at parameters u = 1 and v = 1.

fn bounding_cone_with_origin(&self, origin: &Vect) -> (Vect, Scalar)

Computes an infinite bounding cone of this surface.