Struct ncollide2df64​::geom​::Convex​

pub struct Convex {
    // some fields omitted
}

Set of point assumed to form a convex polyline.

Methods

impl Convex

fn new(points: &[Vect]) -> Convex

Creates a polytope from a set of point.

This computes the convex hull of the set of points internally.

fn new_with_margin(points: &[Vect], margin: Scalar) -> Convex

Creates a polytope from a set of point.

This computes the convex hull of the set of points internally.

unsafe fn new_with_convex_polyline(mesh: Polyline<Scalar, Vect>, margin: Scalar) -> Convex

Creates a polytope from a convex polyline. The convexity is not checked.

fn mesh<'a>(&'a self) -> &'a Polyline<Scalar, Vect>

The convex polyline of this geometry.

fn mesh_mut<'a>(&'a mut self) -> &'a mut Polyline<Scalar, Vect>

The mutable convex polyline of this geometry.

fn unwrap(self) -> Polyline<Scalar, Vect>

The convex polyline of this geometry.

impl Convex

fn pts<'a>(&'a self) -> &'a [Vect]

The list of points of this convex polytope.

fn margin(&self) -> Scalar

The margin surrounding this convex polytope.