Struct ncollide2df64​::geom​::Plane​

pub struct Plane {
    // some fields omitted
}

Implicit description of a plane.

Methods

impl Plane

fn new(normal: Vect) -> Plane

Builds a new plane from its center and its normal.

unsafe fn new_normalized(normal: Vect) -> Plane

Builds a new plane from its center and its normal.

impl Plane

fn normal(&self) -> Vect

The plane normal.