Module ncollide2df64​::narrow​

Narrow phases.

Modules

algorithm
Algorithms needed for distance and penetration depth computation.

closest_points
Functions to compute the closest points between two geometries.

collide
Functions to compute one contact point between two geometries.

surface_selector
Heuristics to select surface containing the projection of a point.

surface_subdivision_tree
Tree used to cache subdivisions of surfaces.

toi
Functions to compute the time of impact between two geometries.

Structs

BallBall
Collision detector between two balls.

BallBezierSurface
Collision detector between a ball and a bezier surface.

BezierSurfaceBall
Collision detector between a bezier surface and a ball.

BezierSurfaceBezierSurface
A collision detector between two bézier surfaces.

ConcaveGeomGeom
Collision detector between a concave geometry and another geometry.

ConcaveGeomGeomFactory
Structure implementing CollisionDetectorFactory in order to create a new ConcaveGeomGeom collision detector.

Contact
Geometric description of a contact.

Empty
A collision detector that does nothing.

GeomConcaveGeom
Collision detector between a geometry and a concave geometry.

GeomConcaveGeomFactory
Structure implementing CollisionDetectorFactory in order to create a new GeomConcaveGeom collision detector.

GeomGeomDispatcher
Collision dispatcher between two ~Geom.

ImplicitImplicit
Persistent collision detector between two shapes having a support mapping function.

ImplicitPlane
Collision detector between a plane and a geometry implementing the Implicit trait.

IncrementalContactManifoldGenerator
Contact manifold generator which keeps track of several contacts.

OneShotContactManifoldGenerator
Contact manifold generator producing a full manifold at the first update.

PlaneImplicit
Collision detector between a plane and a geometry implementing the Implicit trait.

Traits

CollisionDetector
Trait of the algorithms executed during the so-called Narrow Phase.

CollisionDetectorFactory
Trait of structures able do build a new collision detector.

DynamicCollisionDetector
Trait to be implemented by collision detector using dynamic dispatch.

GeomGeomCollisionDetector
Same as the CollisionDetector trait but using dynamic dispatch on the geometries.