Module ncollide2df64​::narrow​::algorithm​::gjk​

The Gilbert–Johnson–Keerthi distance algorithm.

Enums

GJKResult
Results of the GJK algorithm.

Functions

closest_points
Computes the closest points between two convex geometries unsing the GJK algorithm.

closest_points_without_margin
Computes the closest points between two convex geometries without their margins unsing the GJK algorithm.

closest_points_without_margin_with_max_dist
Computes the closest points between two convex geometries without their margins unsing the GJK algorithm.

project_origin
Projects the origin on a geometry unsing the GJK algorithm.

project_origin_with_max_dist
Projects the origin on a geometry using the Separating Axis GJK algorithm. The algorithm will stop as soon as the polytope can be proven to be at least max_dist away from the origin.