Enum ncollide2df64​::narrow​::algorithm​::gjk​::GJKResult​

pub enum GJKResult<_V, Dir> {
    Intersection,
    Projection(_V),
    NoIntersection(Dir),
}

Results of the GJK algorithm.

Variants

Intersection

Result of the GJK algorithm when the origin is inside of the polytope.

Projection

Result of the GJK algorithm when a projection of the origin on the polytope is found.

NoIntersection

Result of the GJK algorithm when the origin is to far away from the polytope.