Function ncollide2df64​::narrow​::algorithm​::minkowski_sampling​::closest_points​

pub fn closest_points<S: Simplex<AnnotatedPoint>, G1: Implicit<Vect, Matrix> + PreferedSamplingDirections<Vect, Matrix>, G2: Implicit<Vect, Matrix> + PreferedSamplingDirections<Vect, Matrix>>(m1: &Matrix, g1: &G1, m2: &Matrix, g2: &G2, simplex: &mut S) -> Option<(Vect, Vect)>

Computes the closest points between two implicit inter-penetrating shapes. Returns None if the shapes are not in penetration. This can be used as a fallback algorithm for the GJK algorithm.