Function ncollide2df64​::narrow​::collide​::implicit_implicit​

pub fn implicit_implicit<S: Simplex<AnnotatedPoint>, G1: Implicit<Vect, Matrix> + PreferedSamplingDirections<Vect, Matrix>, G2: Implicit<Vect, Matrix> + PreferedSamplingDirections<Vect, Matrix>>(m1: &Matrix, g1: &G1, m2: &Matrix, g2: &G2, prediction: &Scalar, simplex: &mut S, init_dir: Option<Vect>) -> GJKResult<Contact, Vect>

Computes a contact point between two implicit geometries.

For optimizations purposes the objects are artificially enlarged by a small margin. This uses the GJK algorithm to handle penetrations within the margin and a Minkowski sum sampling based algorithm to handle deep penetrations.

Arguments:

g1 - the first implicit shape involved on the collision check

g2 - the second implicit shape involved on the collision check

simplex - the simplex the GJK algorithm must use. It is reinitialized before being passed to GJK.