Struct ncollide2df64​::narrow​::BallBezierSurface​

pub struct BallBezierSurface<S, D> {
    // some fields omitted
}

Collision detector between a ball and a bezier surface.

Methods

impl<S: SurfaceSelector<D>, D> BallBezierSurface<S, D>

fn new(selector: S, prediction: Scalar, cache: Arc<RWLock<SurfaceSubdivisionTreeCache<D>>>) -> BallBezierSurface<S, D>

Creates a new collision detector with the given prediction margin.