Struct ncollide2df64​::narrow​::PlaneImplicit​

pub struct PlaneImplicit<G> {
    // some fields omitted
}

Collision detector between a plane and a geometry implementing the Implicit trait.

This detector generates only one contact point. For a full manifold generation, see IncrementalContactManifoldGenerator.

Methods

impl<G> PlaneImplicit<G>

fn new(prediction: Scalar) -> PlaneImplicit<G>

Creates a new persistent collision detector between a plane and a geometry with a support mapping function.