Struct ncollide2df64​::procedural​::path​::ArrowheadCap​

pub struct ArrowheadCap<N> {
    // some fields omitted
}

A cap that looks like an arrow.

Methods

impl<N: Clone + Cast<f64> + FloatMath + Vec3MulRhs<N, Vec3<N>> + Vec3DivRhs<N, Vec3<N>>> ArrowheadCap<N>

fn new(radius_scale: N, front_dist_to_head: N, back_dist_to_head: N) -> ArrowheadCap<N>

Creates a cap that looks like an arrow.

Arguments:

radius_scale - scale factor of the cap base.

front_dist_to_head - distance from the path endpoint and the arrow tip.

back_dist_to_head - distance from the path endpoint and the cap base.