Struct ncollide2df64​::procedural​::path​::PolylinePath​

pub struct PolylinePath<'a, N, V> {
    // some fields omitted
}

A path with its sample points given by a polyline.

This will return sequencially each vertex of the polyline.

Methods

impl<'a, N: Float, V: FloatVec<N> + Clone> PolylinePath<'a, N, V>

fn new(polyline: &'a Polyline<N, V>) -> PolylinePath<'a, N, V>

Creates a new polyline-based path.