Trait ncollide2df64​::ray​::RayCast​

pub trait RayCast {
    fn toi_and_normal_with_ray(&self, ray: &Ray, solid: bool) -> Option<RayIntersection>;

    fn toi_with_ray(&self, ray: &Ray, solid: bool) -> Option<Scalar> { ... }
    fn toi_and_normal_and_uv_with_ray(&self, ray: &Ray, solid: bool) -> Option<RayIntersection> { ... }
    fn intersects_ray(&self, ray: &Ray) -> bool { ... }
    fn toi_with_transform_and_ray(&self, m: &Matrix, ray: &Ray, solid: bool) -> Option<Scalar> { ... }
    fn toi_and_normal_with_transform_and_ray(&self, m: &Matrix, ray: &Ray, solid: bool) -> Option<RayIntersection> { ... }
    fn toi_and_normal_and_uv_with_transform_and_ray(&self, m: &Matrix, ray: &Ray, solid: bool) -> Option<RayIntersection> { ... }
    fn intersects_with_transform_and_ray(&self, m: &Matrix, ray: &Ray) -> bool { ... }
}

Traits of objects which can be tested for intersection with a ray.

Required Methods

fn toi_and_normal_with_ray(&self, ray: &Ray, solid: bool) -> Option<RayIntersection>

Computes the intersection point between this geometry and a ray.

Provided Methods

fn toi_with_ray(&self, ray: &Ray, solid: bool) -> Option<Scalar>

Computes the time of impact between this geometry and a ray

fn toi_and_normal_and_uv_with_ray(&self, ray: &Ray, solid: bool) -> Option<RayIntersection>

Computes the intersection point and normal between this geometry and a ray.

fn intersects_ray(&self, ray: &Ray) -> bool

Tests whether a ray intersects this geometry.

fn toi_with_transform_and_ray(&self, m: &Matrix, ray: &Ray, solid: bool) -> Option<Scalar>

Computes the time of impact between this transform geometry and a ray.

fn toi_and_normal_with_transform_and_ray(&self, m: &Matrix, ray: &Ray, solid: bool) -> Option<RayIntersection>

Computes the time of impact, and normal between this transformed geometry and a ray.

fn toi_and_normal_and_uv_with_transform_and_ray(&self, m: &Matrix, ray: &Ray, solid: bool) -> Option<RayIntersection>

Computes time of impact, normal, and texture coordinates (uv) between this transformed geometry and a ray.

fn intersects_with_transform_and_ray(&self, m: &Matrix, ray: &Ray) -> bool

Tests whether a ray intersects this transformed geometry.